schwab_requirements.txt
# Schwab API Dependencies
# Install with: pip install -r schwab_requirements.txt

# Core HTTP & OAuth
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Optional: For async requests (future enhancement)
# aiohttp>=3.8.0

# Optional: For database storage
# sqlalchemy>=2.0.0

